ui: decouple pan + zoom from panning in "scrollTo"
This CL separates out "focusing" (which means to pan + zoom) from
"panning" (which means just pan, no zoom). We were putting the two in
the same bucket which was a bad idea as in many contexts you only want
panning but not zooming. But in the 'F' keyboard shortcut case, you
*actually* do want to zoom in.