ui: decouple pan + zoom from panning in "scrollTo" (#3831)
This CL separates out "focusing" (which means to pan + zoom) from
"panning" (which means just pan, no zoom)

The motivation for this change:
1) Before https://github.com/google/perfetto/pull/3761, focus used to
work in a very janky way where pressing F once would cause pan to center
the slice and pressing again would zoom in
2) After https://github.com/google/perfetto/pull/3761, F consistently
pans and zooms into the slice which is selected. But it accidentally
also changed the behaviour of a bunch of other places which used focus
to *just* pan.
3) After this change, we change things so that *only* F pans and zooms
and all other places just pan which is what we expect most of the time.
We can also change other places if it becomes necessary to opt into this
new behaviour.