ui: Make checkbox and switch widgets flow inline with text (#3820)
This patch makes the Switch and Checkbox widgets flow inline with
surrounding text. This means the baseline of the labels can be aligned
vertically with surrounding text properly, similar to buttons and other
widgets, and also the label text will wrap properly if placed in a
normal flow of text.

While in the area, combine the `Switch` and `Checkbox` widgets into one
to avoid duplication. The `Checkbox` is the root element with a
`variant` property which can be one of `'switch' | 'checkbox'`, and the
`Switch` component simply wraps the `Checkbox` widget with `variant:
'switch'`.