exp: Cleanup filter node (#3859)
* A redesigned UI for the filter node: no tabbed interface between
freeform SQL and normal filters. Degraded freeform sql to a small button
at the bottom
* Direct filter input: Users can now type simple filters like dur > 1000
directly into a text field, which are then parsed into structured
filters.
* Improved freeform SQL editing: The SQL editor is now in a modal
window, and a truncated version of the query is displayed in the main
view.
* New UI components: ListItem and AdvancedModeChangeButton widgets have
been added to support the new design.
* Enhanced accessibility: Keyboard navigation has been added to the
ListItem component.