exp: Fix slow execution caused by moving data to timeline (#3874)
Moving to timeline requires rerunning the query (preparing the tab) on
each execution. This approach was very costly for explore page, which
creates a lot of executions.

Change it to less costly approach, and later maybe fix the
addQueryResults page to execute the query only when we actually open
timeline.