ui: Ensure focus on events are centered (#3887)
Ensures that all events are centered when using the focus command, covering for various edge cases. For instants, this is when focus is used to zoom to a timeline duration which is less than the minimum timeline duration. For area and slice selections this is when the slice duration is less than the minimum timeline duration.

Bug: #3885