ui: Animate panXIntoView operations (#3922)
All operations that invoke timeline.panXIntoView() are now by default
animated using an ease-in ease-out function that lasts 300ms.

This animation helps the user keep track of where the timeline just
moved to.

If a new animate event is invoked sooner than 500ms after the last, then
we cancel the animation and snap directly to the next position without
animating to avoid the timeline lagging when the user wants to tap
quickly through events.