ui: Improve dataset search performance (#3930)
This patch essentially brings dataset search performance up to the same
standard as the current search implementation.

List of changes:
- Add more information (join and select expressions) to datasets which
is used optimize bulk queries over multiple tracks.
- Generalize the algorithm used to combine multiple datasets but retain
lineage information (required for searching in order to keep track of
which track an event belongs to) + add some unit tests.
- Add an optional search results tab + search benchmarking page to help
with debugging (for development only, for now).