Rebase SQL table on DataGrid
This patch replaces a lot of the custom funcitonality in the SQL table
(Open table... command), with DataGrid, removing a lot of duplicated
code and functionality.

Some of the more custom features have not been replaced, and are still
managed from SQL table, mainly related to the custom cell rendering/
handling such as timestamps, durations and ids, which are out of scope
for DataGrid anyway.

DataGrid now handles:
- Filter UI (menus, distinct values selector)
- Virtual scrolling
- Empty state ('No data', 'No results')
- Sorting menu items
- Fit-to-content menu item

SQLTable still handles:
- Loading data (building queries) via a custom datasource
- Column casting and transforms
- Column adding/hiding
- Custom column types (timestamp, duration, id, ...)

Already handled by Grid (unchangd):
- Column reordering UX
- Virtualization
- Grid layout & column sizing

Change-Id: I69ae4e713808065150617647f5ab5cb017e1f0a0