Refactor `PerfettoCmd::ReportTraceToAndroidFrameworkOrCrash`.
On Android, when uploading traces after reboot, we can have multiple traces,
each of them read from the disk.
To support this, we extract `ReportTraceToAndroidFramework`,
that accepts input arguments, instead of only accepting the
internal ` PerfettoCmd ` state.
It also returns the `base::Status` in case of error instead of crashing
the entire process.

This is a no-op refactoring for the current uploading logic.