ui: DataGrid: Several QoL improvements (#3986)
- Fix broken toolbar - right aligned items were not being aligned to the
right when no filters were present.
- Allow greater customizability of column and cell context menus.
- Move cellRenderer, formatCell to the column level. It makes sense that
they are defined per column rather than globally for the entire grid.
This makes usage a lot cleaner.
- Use the column title instead of id (e.g. 'Wall Duration' instead of
'dur') on filters and in the 'Manage columns' menu.
- Add slice links in query page (switches to timeline and highlights the
slice).
- Add format hints for numeric aggregation values (changes what filters
are displayed) and use this to remove string-like filters in aggregation
panels.
- Make MenuDividers easier to use by hiding the first, last and any
consecutive MenuDividers using CSS.