ui: Grid: Improve column reordering (#4021)
Currently you need to drag and drop a column header over the top of
another column header in order to move the column.

This patch allows dragging the column header over any part of the grid,
including past the last column to put it at the end, which wasn't
supported before.

Plus, the blue column reorder indicator is now rendered down the entire
height of the grid instead of just the header column, so it's much
easier to see.