Speed up the CpuPerUid plugin by 100x. (#3987)
Precompute the total millis per track, and the aggregates "System" and
"Apps" tracks so that the plugin only needs to organize the tracks into
groups. his reduces the plugin load time from 2 seconds to 20
milliseconds.

UI wise, the only change is that the tracks now use the auto-rate
feature, which will show in ms per second, rather than %.