tp: ep: Move Timerange node to SQ (#4101)
The timerange node was generating it's own SQL, which was good for
prototyping but should have proper handling in StructuredQuery.

Created an (experimental) structured query time range, that handles both
dynamic and static cases and migrated the UI to it. Tests are passing
and I didn't notice any regressions.

This was the only blocker to being able to serialize to proto, not json.