[SDK] Expose TrackEvent::IsCategoryEnabledBySession/Config (#4106)
This is an alternative solution to
https://github.com/google/perfetto/pull/4096#issuecomment-3646496843
To enables a common pattern chrome where an observer is interested in a
specific category to control e.g. a timer.

Instead of depending on side-effect of OnStart/OnStop, this PR exposes
IsCategoryEnabledByConfig that can be queried directly on the TrackEvent
session being started/stopped.