stdlib: fix input latency calculation for display frames (#4099)
The `android_input_events` module was missing end-to-end latency data.
This occurred because `surface_frame_token` is NULL (and is never 0) for
display-only frames in the `expected_frame_timeline_slice` table.

Fix the WHERE condition to ensure `end_to_end_latency_dur` is populated
in the input trace.

Bug: 467894962
Test: Enable `com.android.InputEvents` UI plugin and verify input
      latency is visible.
Test: select * from android_input_events;

Change-Id: I979dc5e1d59006ef4501cf737554dbefab5eaa5c

Welcome to Perfetto!
Make sure your PR has a bug/issue attached or has at least
a clear description of the problem you are trying to fix.

For more details please see
https://perfetto.dev/docs/contributing/getting-started