bazel: upgrade to bazel 8 and migrate to fully bzlmod based setup (#4108)
This CL uprevs us to the latest version of Bazel and switches us over to
a full bzlmod setup. bzlmod is the only way to build with Bazel 8,
WORKSPACE based building is not supported and is already bit-rotting in
many places.

Keep the perfetto_deps() fallback for anyone still using workspace based
usage of Perfetto but ideally they would also need to switch to bzlmod
on new versions of Perfetto.