tracing: split out TracingServiceImpl (#4129)
This CL is a pure non-functional split of TracingServiceImpl
into several files. TSI has become too big for humans... and now
even for AI (which starts warning about the file being too large).
This CL is very trivial (I have been Cut/Paste chunks of code) but
is fully human written because I don't yet trust AI copying large
chunks of code without being creative (and the diff would be
impossible for a human to review because of the file moves)

I split it as follows:
- tracing_service_structs: holds all the auxiliary structs
  that are used across different subclasses (e.g. TriggerInfo,
  PendingClone, etc)
- tracing_service_endpoints: the boring proxy layers that
  bridge the lowest-level ext API and calls into real methods
  of TSI. I find these mostly annoying because when I codesearch
  they show up while I am looking for the real impl in TSI.
- tracing_session: it's a pure struct, but it's big enough and
  changed/read enough that warrants its own file.

The only non-copy-paste change I made has been outlining a
few functions in tracing_service_sesison.cc which used to be
inline for no good reasons.