perfetto_cmd: don't delete not empty `write_into_file` trace on error. (#4005)
This bug happens because the `bytes_written_` field was not updated
when the `perfetto_cmd` exits with an error.
To prevent such bugs in the future, we extract the logic to a separate
method that should be called to get the the count of bytes written.

This function is called only at the end, when all the data is written so
there is no need to do something smarter than `fseek`+`ftell`.