sdk: Add string interning support (reland) (#4154)
Introduces support for interning event arguments into specific
InternedData messages. This enables safe handling of app-controlled
strings, starting with AndroidJobName.

The implementation adds a new field type to the C ABI that includes an
|interned_type_id| and updates the C++, Java, and JNI layers of the SDK
to expose this functionality.

This is a reland of https://github.com/google/perfetto/pull/4062 base on
https://github.com/google/perfetto/pull/4081

Bug: 465114453
Test: tools/bazel clean && tools/bazel test
//:src_android_sdk_java_test_perfetto_trace_instrumentation_test
--test_output=streamed