ui: DataGrid: Add support for compound columns (#4162)
Compound columns are columns that depend on multiple fields in the
underlying datasource.

When a column has a dependsOn property, those fields are requested from
the datasource as well as well as the referenced field. This means that
all these fields will be present in the 'row' parameter passed to the
cellRenderer() so that all these fields can be used to format the cell.

Compelling example: build a process column which combines upid,
process.name, and process.pid into one column.

E.g. `[<upid>] <process.name> (<process.pid>)`