Fix bug where standalone query page doesn't update
When running a query in the query page - the tab that is added to the
timeline page is not updated anymore.

This patch adds a new permanent page that can be opened at any time
which simply mirrors the query results displayed in the query page.

This can be closed and reopened at any time so you don't lose anything.

While there, added an 'Add debug track' button to the results on the
query page which adds a debug tracka and also navigates back to the
timeline page.

Change-Id: I224fe08713e86d34439778a94e65fe62919a3468