ui: Fix bug where standalone query tab doesn't update (#4170)
When running a query in the query page - the tab that is added to the
timeline page is not updated anymore.

This patch adds a new permanent page that can be opened at any time
which simply mirrors the query results displayed in the query page.

This can be closed and reopened at any time so you don't lose anything.

While there, added an 'Add debug track' button to the results on the
query page which adds a debug tracka and also navigates back to the
timeline page.

Also rename 'Show debug track' with 'Add debug track' as the operation
is definitely more of an 'add' than a 'show'.