tp: fix parsing of systrace format without irq flags (#4191)
My change in https://github.com/google/perfetto/pull/2778
made it so that traces without IRQ flags would accidentially parse the
start of the timestamp as IRQ flags instead.

Fix this by being stricter with the format of the IRQ flags (i.e. make
it so that everything except preempt depth must not be a digit, make the
entire thing optional, don't allow "partial irq flags").

Fixes: https://github.com/google/perfetto/issues/4190